Fix SELECT timeout on higher cascade levels
`current_level_known_bits` in the select function was being incorrectly
calculated from the uid_index, where on each cascade level it should be
reset to 0. This resulted in an invalid command being sent to the PICC
after the first level. Setting `current_level_known_bits` to 0 on level
increment fixes this and allows Ultralight cards to be identified
correctly.

Note this issue only occurs when more than one cascade level is
required, e.g. on Ultralight cards.